requirements_backup.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.16.1
azure-search-documents==11.4.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dataclasses-json==0.6.7
distro==1.9.0
Flask==3.0.3
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-openai==0.2.2
langchain-text-splitters==0.3.8
langgraph==0.2.76
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.73
langsmith==0.4.6
lxml==5.2.2
MarkupSafe==3.0.2
marshmallow==3.26.1
msal==1.32.3
msal-extensions==1.3.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.1
openai==1.40.8
orjson==3.11.0
ormsgpack==1.10.0
packaging==25.0
propcache==0.3.2
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.10.1
pydantic_core==2.20.1
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
tenacity==9.1.2
tiktoken==0.7.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.2.3
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
